requirements.txt
sqlalchemy==2.0.19
pandas==2.0.3
pyyaml==6.0.1
jinja2==3.1.2
loguru==0.7.0
pytest==7.4.0
